reorganized the folder structure so that cmponents are beter organized. Also fixed the registration form so that users can register correctly and then it gets reidrected to the Profile pags
