+++ client/src/components/layout/Alerts/Alert.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
-import { removeAlert } from "../../actions/alert";
+import { removeAlert } from "../../../actions/alert";
 
 const onSubmit = id => async dispatch => {
   removeAlert(id);
