+++ client/src/components/layout/Spinner.js
@@ -1,14 +1,14 @@
-import React, { Fragment } from 'react';
-import spinner from './spinner.gif';
+import React, { Fragment } from "react";
+import spinner from "../Assets/spinner.gif";
 
 export default () => {
-  return(
-  <Fragment>
-    <img
-      src={spinner}
-      style={{ width: '200px', margin: 'auto', display: 'block' }}
-      alt="Loading..."
-    />
-  </Fragment>
-)
+  return (
+    <Fragment>
+      <img
+        src={spinner}
+        style={{ width: "200px", margin: "auto", display: "block" }}
+        alt="Loading..."
+      />
+    </Fragment>
+  );
 };
